requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
psycopg2-binary
pydantic-settings
python-dotenv
pwdlib[argon2]
PyJWT
email-validator
python-multipart
httpx
pytest
pytest-asyncio
